updated git auto versioning/packaging scheme
now, in order to build .deb package, CMAKE_BUILD_TYPE must be package.
Additionally, in order to get a nicely versioned package, the currently
checked out commit must be sitting directly atop an annotated tag of the
form v1.0.0, and the working directory must be clean. Otherwise, the
version will be the current commit hash, and -dirty is appended if the
work area is not clean