don't try to determine primary sample idx when not merging samples
Entry merging requires figuring out which entry should override another
in the case of collisions. This only happens when merging samples,
however. The test to see which entry is better only makes sense when
you have the same sample in multiple input files, and it is expensive.
Now we don't do it when we're just doing a simple merge with no sample
collisions.