change < to <= when extending reference alleles
in the AlleleMerger, when merging reference alleles from multiple
entries, we don't need to do anything when an entry is completely
contained in the current merged ref allele. We were comparing with <
rather than <=. This doesn't change output, it just keeps us from
running code like "append the empty string to the reference allele".